Move third-party modules into a standalone test module.
